refactor(users): rename from user to users
