--- apps/api/src/users/user.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UserService } from './user.service';
-import { UserResolver } from './user.resolver';
-import { DatabaseModule } from 'src/database/database.module';
-
-@Module({
-  imports: [DatabaseModule],
-  providers: [UserService, UserResolver],
-})
-export class UserModule {}
